# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-tenant SaaS API",
    "description": "API for the Multi-tenant SaaS application (DB schema-aligned)",
    "version": "0.0.1"
  },
  "paths": {
...
    "/v1/organizations/": {
      "post": {
        "tags": ["organizations"],
        "summary": "Create Organization",
        "description": "Create a new organization in org.organizations.",
        "operationId": "create_organization_v1_organizations__post",
        "security": [{ "HTTPBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrganizationCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organization created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrganizationPublic" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["organizations"],
        "summary": "List Organizations",
        "description": "List organizations from org.organizations.",
        "operationId": "list_organizations_v1_organizations__get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of organizations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/OrganizationPublic" }
                }
              }
            }
          }
        }
      }
    },
    "/v1/organizations/{organization_id}": {
      "get": {
        "tags": ["organizations"],
        "summary": "Get Organization",
        "description": "Get an organization by ID from org.organizations.",
        "operationId": "get_organization_v1_organizations__organization_id__get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Organization Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrganizationPublic" }
              }
            }
          },
          "404": {
            "description": "Organization not found"
          }
        }
      },
      "put": {
        "tags": ["organizations"],
        "summary": "Update Organization",
        "description": "Update an organization in org.organizations.",
        "operationId": "update_organization_v1_organizations__organization_id__put",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Organization Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrganizationUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organization updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrganizationPublic" }
              }
            }
          },
          "404": {
            "description": "Organization not found"
          }
        }
      },
      "delete": {
        "tags": ["organizations"],
        "summary": "Soft Delete Organization",
        "description": "Soft delete an organization (sets deleted_at) in org.organizations.",
        "operationId": "delete_organization_v1_organizations__organization_id__delete",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Organization Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization soft deleted"
          },
          "404": {
            "description": "Organization not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrganizationCreate": {
        "type": "object",
        "required": ["name", "slug"],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 50,
            "title": "Slug"
          },
          "avatar_url": {
            "type": "string",
            "maxLength": 512,
            "title": "Avatar URL"
          },
          "is_active": {
            "type": "boolean",
            "default": true,
            "title": "Is Active"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          }
        },
        "title": "OrganizationCreate",
        "description": "Schema for creating organizations in org.organizations."
      },
      "OrganizationPublic": {
        "type": "object",
        "required": ["id", "name", "slug", "is_active", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "slug": { "type": "string", "title": "Slug" },
          "avatar_url": { "type": "string", "title": "Avatar URL" },
          "is_active": { "type": "boolean", "title": "Is Active" },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Deleted At"
          }
        },
        "title": "OrganizationPublic",
        "description": "Public organization schema for API responses from org.organizations."
      },
      "OrganizationUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 50,
            "title": "Slug"
          },
          "avatar_url": {
            "type": "string",
            "maxLength": 512,
            "title": "Avatar URL"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          }
        },
        "title": "OrganizationUpdate",
        "description": "Schema for updating organizations in org.organizations."
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}